Repo cleaning, README update, and better environment names
