+++ custom_envs/envs/__init__.py
@@ -1,0 +1,3 @@
+from custom_envs.envs.push_box import PushBoxBaseEnv
+from custom_envs.envs.push_box_random import PushBoxRandEnv
+from custom_envs.envs.push_box_random_2 import PushBoxRandEnv2
