+++ custom_envs/envs/push_box_random.py
@@ -5,11 +5,14 @@ import gymnasium as gym
 from gymnasium import spaces
 
 
-class GridWorldEnv(gym.Env):
+class PushBoxRandEnv(gym.Env):
     metadata = {"render_modes": ["human", "rgb_array"], "render_fps": 4}
 
-    def __init__(self, render_mode=None, size=5):
+    def __init__(self, render_mode=None, size=5, step_limit=None):
         self.size = size  # The size of the square grid
+        # steps before target changes position
+        if step_limit is None: self.step_limit = size
+        else: self.step_limit = step_limit
         self.window_size = 512  # The size of the PyGame window
 
         # Observations are dictionaries with the agent's and the target's location.
@@ -21,8 +24,8 @@ class GridWorldEnv(gym.Env):
             }
         )
 
-        # We have 4 actions, corresponding to "right", "up", "left", "down", "right"
-        self.action_space = spaces.Discrete(4)
+        # We have 4 actions, corresponding to "right", "up", "left", "down", "right", "stay"
+        self.action_space = spaces.Discrete(5)
 
         """
         The following dictionary maps abstract actions from `self.action_space` to 
@@ -34,6 +37,7 @@ class GridWorldEnv(gym.Env):
             1: np.array([0, 1]),
             2: np.array([-1, 0]),
             3: np.array([0, -1]),
+            4: np.array([0, 0]),
         }
 
         assert render_mode is None or render_mode in self.metadata["render_modes"]
@@ -63,6 +67,8 @@ class GridWorldEnv(gym.Env):
         # We need the following line to seed self.np_random
         super().reset(seed=seed)
 
+        self.step_limit = self.size # steps before target changes position
+
         # Choose the agent's location uniformly at random
         self._agent_location = self.np_random.integers(0, self.size, size=2, dtype=int)
 
@@ -85,8 +91,17 @@ class GridWorldEnv(gym.Env):
     def step(self, action):
         # An episode is done iff the agent has reached the target
         terminated = self._target_location[0] == self.size - 1
+
+        # if self.step_limit == 0 change target location
+        if self.step_limit == 0:
+            self.step_limit = self.size
+            self._target_location = self._agent_location
+            while np.array_equal(self._target_location, self._agent_location) or self._target_location[0] in [0, self.size-1]:
+                self._target_location = self.np_random.integers(
+                    0, self.size, size=2, dtype=int
+            )
         
-        # Map the action (element of {0,1,2,3}) to the direction we walk in
+        # Map the action (element of {0,1,2,3,4}) to the direction we walk in
         direction = self._action_to_direction[action]
         # We use `np.clip` to make sure we don't leave the grid
         if np.array_equal(self._agent_location + direction, self._target_location):
@@ -111,6 +126,11 @@ class GridWorldEnv(gym.Env):
             self._agent_location = np.clip(
                 self._agent_location + direction, 0, self.size - 1
             )
+
+        # reduce step limit
+        self.step_limit -= 1
+
+
         # An episode is truncated iff the agent has reached the left border of the grid
         truncated = self._target_location[0] == 0
         reward = 1 if terminated else 0  # Binary sparse rewards
@@ -140,10 +160,10 @@ class GridWorldEnv(gym.Env):
             self.window_size / self.size
         )  # The size of a single grid square in pixels
 
-        # First, we draw the green goal zone
+        # First, we draw the orange goal zone
         pygame.draw.rect(
             canvas,
-            (0, 255, 0),
+            (255, 165, 0),
             pygame.Rect(
                 (pix_square_size * (self.size-1), 0),
                 (pix_square_size, pix_square_size*self.size),
@@ -153,7 +173,7 @@ class GridWorldEnv(gym.Env):
         # we draw the target
         pygame.draw.rect(
             canvas,
-            (255, 0, 0),
+            (238, 130, 238),
             pygame.Rect(
                 pix_square_size * self._target_location,
                 (pix_square_size, pix_square_size),
@@ -162,7 +182,7 @@ class GridWorldEnv(gym.Env):
         # Now we draw the agent
         pygame.draw.circle(
             canvas,
-            (0, 0, 255),
+            (64, 224, 208),
             (self._agent_location + 0.5) * pix_square_size,
             pix_square_size / 3,
         )
