--- gym_examples/envs/__init__.py
@@ -1,3 +1,0 @@
-from gym_examples.envs.grid_world import GridWorldEnv
-from gym_examples.envs.grid_world_random import GridWorldRandEnv
-from gym_examples.envs.grid_world_random2 import GridWorldRandEnv2
