--- gym_examples/wrappers/__init__.py
@@ -1,4 +1,0 @@
-from gym_examples.wrappers.clip_reward import ClipReward
-from gym_examples.wrappers.discrete_actions import DiscreteActions
-from gym_examples.wrappers.reacher_weighted_reward import ReacherRewardWrapper
-from gym_examples.wrappers.relative_position import RelativePosition
