+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup
 
 setup(
-    name="gym_examples",
+    name="custom_envs",
     version="0.0.1",
     install_requires=["gymnasium==0.28.1", "pygame==2.1.0"],
 )
